@@ -30,7 +30,7 @@ public struct GoogleError: Error, Decodable {
 public struct FCMError: Error, Decodable {
     public let errorCode: ErrorCode
 
-    public enum ErrorCode: String, Decodable {
+    public enum ErrorCode: String, Decodable, Sendable {
         case unspecified = "UNSPECIFIED_ERROR"
         case invalid = "INVALID_ARGUMENT"
         case unregistered = "UNREGISTERED"
@@ -45,15 +45,20 @@ public struct FCMError: Error, Decodable {
 extension EventLoopFuture where Value == ClientResponse {
     func validate() -> EventLoopFuture<ClientResponse> {
         return flatMapThrowing { (response) in
-            guard 200 ..< 300 ~= response.status.code else {
-                if let error = try? response.content.decode(GoogleError.self) {
-                    throw error
-                }
-                let body = response.body.map(String.init) ?? ""
-                throw Abort(.internalServerError, reason: "FCM: Unexpected error '\(body)'")
-            }
-
+            try response.validate()
             return response
+        }
+    }
+}
+
+extension ClientResponse {
+    func validate() throws {
+        guard 200 ..< 300 ~= self.status.code else {
+            if let error = try? self.content.decode(GoogleError.self) {
+                throw error
+            }
+            let body = self.body.map(String.init) ?? ""
+            throw Abort(.internalServerError, reason: "FCM: Unexpected error '\(body)'")
         }
     }
 }
