@@ -2,23 +2,23 @@ import Foundation
 import Vapor
 
 extension FCM {
-    public func createTopic(_ name: String? = nil, tokens: String...) -> EventLoopFuture<String> {
-        createTopic(name, tokens: tokens)
+//    public func createTopic(_ name: String? = nil, tokens: String...) async throws -> String {
+//        try await createTopic(name, tokens: tokens)
+//    }
+//
+//    public func createTopic(_ name: String? = nil, tokens: String..., on eventLoop: EventLoop) async throws -> String {
+//        try await createTopic(name, tokens: tokens).hop(to: eventLoop)
+//    }
+
+    public func createTopic(_ name: String? = nil, tokens: [String]) async throws -> String {
+        try await _createTopic(name, tokens: tokens)
     }
 
-    public func createTopic(_ name: String? = nil, tokens: String..., on eventLoop: EventLoop) -> EventLoopFuture<String> {
-        createTopic(name, tokens: tokens).hop(to: eventLoop)
-    }
+//    public func createTopic(_ name: String? = nil, tokens: [String], on eventLoop: EventLoop) async throws -> String {
+//        try await _createTopic(name, tokens: tokens).hop(to: eventLoop)
+//    }
 
-    public func createTopic(_ name: String? = nil, tokens: [String]) -> EventLoopFuture<String> {
-        _createTopic(name, tokens: tokens)
-    }
-
-    public func createTopic(_ name: String? = nil, tokens: [String], on eventLoop: EventLoop) -> EventLoopFuture<String> {
-        _createTopic(name, tokens: tokens).hop(to: eventLoop)
-    }
-
-    private func _createTopic(_ name: String? = nil, tokens: [String]) -> EventLoopFuture<String> {
+    private func _createTopic(_ name: String? = nil, tokens: [String]) async throws -> String {
         guard let configuration = self.configuration else {
             fatalError("FCM not configured. Use app.fcm.configuration = ...")
         }
@@ -27,27 +27,33 @@ extension FCM {
         }
         let url = self.iidURL + "batchAdd"
         let name = name ?? UUID().uuidString
-        return getAccessToken().flatMap { accessToken -> EventLoopFuture<ClientResponse> in
-            var headers = HTTPHeaders()
-            headers.add(name: .authorization, value: "key=\(serverKey)")
+        
+        let _ = try await getAccessToken()
+        var headers = HTTPHeaders()
+        headers.add(name: .authorization, value: "key=\(serverKey)")
 
-            return self.client.post(URI(string: url), headers: headers) { (req) in
-                struct Payload: Content {
-                    let to: String
-                    let registration_tokens: [String]
-                    
-                    init(to: String, registration_tokens: [String]) {
-                        self.to = "/topics/\(to)"
-                        self.registration_tokens = registration_tokens
-                    }
+        let response = try await self.client.post(URI(string: url), headers: headers) { (req) in
+            struct Payload: Content {
+                let to: String
+                let registration_tokens: [String]
+                
+                init(to: String, registration_tokens: [String]) {
+                    self.to = "/topics/\(to)"
+                    self.registration_tokens = registration_tokens
                 }
-                let payload = Payload(to: name, registration_tokens: tokens)
-                try req.content.encode(payload)
             }
+            let payload = Payload(to: name, registration_tokens: tokens)
+            try req.content.encode(payload)
         }
-        .validate()
-        .map { _ in
-            return name
+        
+        guard 200 ..< 300 ~= response.status.code else {
+            if let error = try? response.content.decode(GoogleError.self) {
+                throw error
+            }
+            let body = response.body.map(String.init) ?? ""
+            throw Abort(.internalServerError, reason: "FCM: Unexpected error '\(body)'")
         }
+        
+        return name
     }
 }
