@@ -2,7 +2,7 @@ import Foundation
 import JWT
 
 extension FCM {
-    func generateJWT() throws -> String {
+    func generateJWT() async throws -> String {
         guard let configuration = self.configuration else {
             fatalError("FCM not configured. Use app.fcm.configuration = ...")
         }
@@ -14,17 +14,17 @@ extension FCM {
         }
         gAuth = gAuth.updated()
         self.gAuth = gAuth
-        let pk = try RSAKey.private(pem: pemData)
-        let signer = JWTSigner.rs256(key: pk)
-        return try signer.sign(gAuth)
+        let pk = try Insecure.RSA.PrivateKey(pem: pemData)
+        let keys = await JWTKeyCollection().add(rsa: pk, digestAlgorithm: .sha256)
+        return try await keys.sign(gAuth)
     }
     
-    func getJWT() throws -> String {
+    func getJWT() async throws -> String {
         guard let gAuth = gAuth else { fatalError("FCM gAuth can't be nil") }
         if !gAuth.hasExpired, let jwt = jwt {
             return jwt
         }
-        let jwt = try generateJWT()
+        let jwt = try await generateJWT()
         self.jwt = jwt
         return jwt
     }
