@@ -2,15 +2,11 @@ import Foundation
 import Vapor
 
 extension FCM {
-    public func send(_ message: FCMMessageDefault) -> EventLoopFuture<String> {
-        _send(message)
+    public func send(_ message: FCMMessageDefault) async throws -> String {
+        try await _send(message)
     }
     
-    public func send(_ message: FCMMessageDefault, on eventLoop: EventLoop) -> EventLoopFuture<String> {
-        _send(message).hop(to: eventLoop)
-    }
-    
-    private func _send(_ message: FCMMessageDefault) -> EventLoopFuture<String> {
+    private func _send(_ message: FCMMessageDefault) async throws -> String {
         guard let configuration = self.configuration else {
             fatalError("FCM not configured. Use app.fcm.configuration = ...")
         }
@@ -28,25 +24,25 @@ extension FCM {
         }
 
         let url = actionsBaseURL + configuration.projectId + "/messages:send"
-        return getAccessToken().flatMap { accessToken -> EventLoopFuture<ClientResponse> in
-            var headers = HTTPHeaders()
-            headers.bearerAuthorization = .init(token: accessToken)
-
-            return self.client.post(URI(string: url), headers: headers) { (req) in
-                struct Payload: Content {
-                    let message: FCMMessageDefault
-                }
-                let payload = Payload(message: message)
-                try req.content.encode(payload)
+        
+        let accessToken = try await getAccessToken()
+        var headers = HTTPHeaders()
+        headers.bearerAuthorization = .init(token: accessToken)
+        
+        let response = try await self.client.post(URI(string: url), headers: headers) { (req) in
+            struct Payload: Content {
+                let message: FCMMessageDefault
             }
+            let payload = Payload(message: message)
+            try req.content.encode(payload)
         }
-        .validate()
-        .flatMapThrowing { res in
-            struct Result: Decodable {
-                let name: String
-            }
-            let result = try res.content.decode(Result.self)
-            return result.name
+        
+        try response.validate()
+        
+        struct Result: Decodable {
+            let name: String
         }
+        let result = try response.content.decode(Result.self)
+        return result.name
     }
 }
